Fix documentation restricted Google Sheets query
When querying a restricted access Google Sheets, having the Google Service Account as a part of the organization it's restricted to is not enough to have read access to the sheets. The email address of the Service Account needs to be explicitly added to the "Who Can Access" list to allow it to read.