WIP: Scaffolds in most of the information I need.
There are a bunch of pictures waiting on the side to be edited here.

The gif was made in FCPX and rendered using ffmpeg. It would be nice
to have a gallery component in our Gatsby site so I could just use a
slide show. I've got the images tucked away for just this reason in
case a gallery component is available sometime in the future.